+++ examples/basic/config/config.exs
@@ -3,4 +3,7 @@ import Config
 config :logger,
   level: :warning
 
+config :snakepit,
+  auto_install_python_deps: true
+
 # Snakepit is configured in runtime.exs using SnakeBridge.ConfigHelper