+++ examples/coverage_report_example/config/config.exs
@@ -8,3 +8,6 @@ config :snakebridge,
 
 config :logger,
   level: :warning
+
+config :snakepit,
+  auto_install_python_deps: true