+++ examples/proof_pipeline/config/config.exs
@@ -9,4 +9,7 @@ config :snakebridge,
 config :logger,
   level: :warning
 
+config :snakepit,
+  auto_install_python_deps: true
+
 # Snakepit is configured in runtime.exs using SnakeBridge.ConfigHelper